Changes - synching data with server, bug fixes
Fixed spelling mistake in GPS alert text.
Slightly increased font size and padding of text in cards dialog in Preferences screen
Modified the way soft keyboard is shown in search screen
Every time app is launched (not resumed), it checks the server for change in data. If version changes, it downloads and extracts latest zip from server
App is bundled with db package, which is extracted and used at first time launch - so even if internet is not available, the user will have the data.